add seeder auth

+++ app/Http/Controllers/Backsite/SpecialistController.php
@@ -77,6 +77,7 @@ class SpecialistController extends Controller
      */
     public function show(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         // menggunakan fitur model binding, Models dijadikan sebuah parameter pada function
         return view('pages.backsite.master-data.specialist.show', compact('specialist'));
     }
@@ -89,6 +90,8 @@ class SpecialistController extends Controller
      */
     public function edit(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('pages.backsite.master-data.specialist.edit', compact('specialist'));
     }
 
@@ -120,6 +123,8 @@ class SpecialistController extends Controller
      */
     public function destroy(Specialist $specialist)
     {
+        abort_if(Gate::denies('specialist_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $specialist->forceDelete();
 
         alert()->success('Success Message', 'Successfully deleted specialist');
